Change generation of filenames #3427
